requirements.txt
django
channels
vjoy